@@ -1,6 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: Copyright The Lance Authors
 
+use std::sync::LazyLock;
 use std::time::Duration;
 
 use crate::Result;
@@ -31,18 +32,23 @@ pub fn get_num_compute_intensive_cpus() -> usize {
     num_cpus::get() - *IO_CORE_RESERVATION
 }
 
-lazy_static::lazy_static! {
-    pub static ref IO_CORE_RESERVATION: usize = std::env::var("LANCE_IO_CORE_RESERVATION").unwrap_or("2".to_string()).parse().unwrap();
+pub static IO_CORE_RESERVATION: LazyLock<usize> = LazyLock::new(|| {
+    std::env::var("LANCE_IO_CORE_RESERVATION")
+        .unwrap_or("2".to_string())
+        .parse()
+        .unwrap()
+});
 
-    pub static ref CPU_RUNTIME: Runtime = Builder::new_multi_thread()
+pub static CPU_RUNTIME: LazyLock<Runtime> = LazyLock::new(|| {
+    Builder::new_multi_thread()
         .thread_name("lance-cpu")
         .max_blocking_threads(get_num_compute_intensive_cpus())
         .worker_threads(1)
         // keep the thread alive "forever"
         .thread_keep_alive(Duration::from_secs(u64::MAX))
         .build()
-        .unwrap();
-}
+        .unwrap()
+});
 
 /// Spawn a CPU intensive task
 ///
